requirements.txt
telethon
python-dotenv
psycopg2-binary
pandas
sqlalchemy
dbt-postgres
ultralytics
fastapi
uvicorn
dagster
dagster-webserver 
